update with merge
