+++ backend/api/models/user.go
@@ -13,16 +13,36 @@ type User struct {
 	Notifications []Notification `json:"notifications" gorm:"foreignkey:ReceiverId"`
 } //@name User
 
+type UserInfo struct {
+	Id        string `json:"id"`
+	Username  string `json:"username"`
+	Email     string `json:"email"`
+	Status    string `json:"status"`
+	FirstName string `json:"firstName"`
+	LastName  string `json:"lastName"`
+} //@name UserInfo
+
 type UserPublic struct {
 	Id       string `json:"id"`
-	UserName string `json:"userName"`
+	Username string `json:"username"`
 	Status   string `json:"status"`
 } //@name UserPublic
+
+func (user User) ToUserInfoResponse() *UserInfo {
+	return &UserInfo{
+		Id:        user.Id,
+		Username:  user.Username,
+		Status:    user.Status,
+		Email:     user.Email,
+		FirstName: user.FirstName,
+		LastName:  user.LastName,
+	}
+}
 
 func (user User) ToUserResponse() *UserPublic {
 	return &UserPublic{
 		Id:       user.Id,
-		UserName: user.Username,
+		Username: user.Username,
 		Status:   user.Status,
 	}
 }
